% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_diary.R
\name{is_sleep_diary}
\alias{is_sleep_diary}
\title{Test whether an object is a sleep_diary}
\usage{
is_sleep_diary(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a sleep_diary
}
