% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{plot.bland_altman}
\alias{plot.bland_altman}
\title{Plot a Bland-Altman result}
\usage{
\method{plot}{bland_altman}(x, ...)
}
\arguments{
\item{x}{A list returned by \code{bland_altman()}.}

\item{...}{Passed to \code{print.ggplot}.}
}
\value{
The ggplot object, invisibly.
}
\description{
Plot a Bland-Altman result
}
