% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{print.bland_altman}
\alias{print.bland_altman}
\title{Print a Bland-Altman summary}
\usage{
\method{print}{bland_altman}(x, ...)
}
\arguments{
\item{x}{A list returned by \code{bland_altman()}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Bland-Altman summary
}
