% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_diary.R
\name{print.sleep_diary}
\alias{print.sleep_diary}
\title{One-line status of a sleep_diary}
\usage{
\method{print}{sleep_diary}(x, ...)
}
\arguments{
\item{x}{A \code{sleep_diary} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
One-line status of a sleep_diary
}
