% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_diary.R
\name{print.summary.sleep_diary}
\alias{print.summary.sleep_diary}
\title{Print a sleep_diary summary}
\usage{
\method{print}{summary.sleep_diary}(x, ...)
}
\arguments{
\item{x}{A \code{summary.sleep_diary} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a sleep_diary summary
}
