% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_apply_corrections}
\alias{step_apply_corrections}
\title{Step 6 -- apply manual corrections}
\usage{
step_apply_corrections(x, corrections_df, manual_unusual_df)
}
\arguments{
\item{x}{A \code{sleep_diary} object.}

\item{corrections_df}{Data frame of manual error corrections.}

\item{manual_unusual_df}{Data frame of manual unusual-pattern decisions.}
}
\value{
A \code{sleep_diary} object.
}
\description{
Step 6 -- apply manual corrections
}
