% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_process_intervals}
\alias{step_process_intervals}
\title{Step 3 -- parse interval durations}
\usage{
step_process_intervals(
  x,
  vars = c("duration_totalmin_sol_estimate_am", "duration_totalmin_waso_estimate_am",
    "duration_totalmin_napstoday_PM", "exercisetoday_PM_totalmin_Light",
    "exercisetoday_PM_totalmin_Moderate", "exercisetoday_PM_totalmin_Vigorous",
    "exercisetoday_PM_totalmin_Strength")
)
}
\arguments{
\item{x}{A \code{sleep_diary} object.}

\item{vars}{Character vector of interval variables to process.}
}
\value{
A \code{sleep_diary} object.
}
\description{
Step 3 -- parse interval durations
}
