% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_process_timestamps}
\alias{step_process_timestamps}
\title{Step 2 -- parse timestamps}
\usage{
step_process_timestamps(
  x,
  vars = c("time_bed_am", "time_sleep_am", "time_awake_am", "time_getup_am",
    "caffeinetoday_PM", "alcoholtoday_PM", "nicotine_amount_pm", "cannabis_amount_pm")
)
}
\arguments{
\item{x}{A \code{sleep_diary} object.}

\item{vars}{Character vector of timestamp variables to process.}
}
\value{
A \code{sleep_diary} object.
}
\description{
Step 2 -- parse timestamps
}
