% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{run_report}
\alias{run_report}
\title{Run the correction status report}
\usage{
run_report(config = NULL, project_dir = ".")
}
\arguments{
\item{config}{Character or list. Config path or config list.}

\item{project_dir}{Character. Path to project root.}
}
\description{
Run the correction status report
}
